refactor(ui): move shared-table into ui/datatable, drop barrel
Relocate src/components/shared-table to src/components/ui/datatable now that the
vendored ui/ folder is the home for shared primitives. Remove the index.tsx
barrel (it only re-exported MetricCell/EMPTY_METRIC/MetricHeaderCell); importers
reference the source files directly.
