+++ src/components/proxmox/ProxmoxHostView.tsx
@@ -4,7 +4,7 @@ import { Collapsible, CollapsibleContent } from '@/components/ui/collapsible'
 import { ChevronRight, Server } from 'lucide-react'
 import type { ProxmoxClusterOverview, GuestRow } from '@/types/proxmox'
 import { useGeneralSettings, useProxmoxSettings } from '@/hooks/useSettings'
-import { EMPTY_METRIC } from '@/components/shared-table'
+import { EMPTY_METRIC } from '@/components/ui/datatable/MetricCell'
 import { formatUptime } from '@/components/proxmox/utils'
 import { GuestSection } from '@/components/proxmox/GuestSection'
 import { StorageSection } from '@/components/proxmox/StorageSection'
