--- src/components/shared-table/index.tsx
@@ -1,2 +1,0 @@
-export { MetricCell, EMPTY_METRIC } from '@/components/shared-table/MetricCell';
-export { MetricHeaderCell } from '@/components/shared-table/MetricHeaderCell';
