+++ src/components/ui/datatable/MetricCell.tsx
@@ -1,5 +1,5 @@
 import { memo, type ReactNode } from 'react';
-import SparklineCell from '@/components/shared-table/SparklineCell';
+import SparklineCell from '@/components/ui/datatable/SparklineCell';
 import { abbreviateUnit } from '@/lib/utils/abbreviate-unit';
 
 /** Display placeholder for metrics with no value */
