+++ src/components/ui/datatable/SparklineCanvas.tsx
