+++ src/components/zfs/subtables/PoolSubTable.tsx
@@ -1,6 +1,6 @@
 import { memo, useCallback, useMemo, type ReactNode } from 'react';
 import type { ColumnDef, ExpandedState } from '@tanstack/react-table';
-import { DataTable } from '@/components/shared-table/DataTable';
+import { DataTable } from '@/components/ui/datatable/DataTable';
 import type { ZFSTableRow } from '@/components/zfs/ZFSPoolsTable';
 import VdevDiskSubTable from '@/components/zfs/subtables/VdevDiskSubTable';
