+++ src/components/zfs/subtables/VdevDiskSubTable.tsx
@@ -1,6 +1,6 @@
 import { memo, useCallback, useMemo, type ReactNode } from 'react';
 import type { ColumnDef, ExpandedState } from '@tanstack/react-table';
-import { DataTable } from '@/components/shared-table/DataTable';
+import { DataTable } from '@/components/ui/datatable/DataTable';
 import type { ZFSTableRow } from '@/components/zfs/ZFSPoolsTable';
 import DiskSubTable from '@/components/zfs/subtables/DiskSubTable';
